Fix midi playback on 64 bit sims and const a table
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@14842 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/midi/midiplay.c
@@ -92,7 +92,7 @@ struct MIDIfile * mf IBSS_ATTR;
 int numberOfSamples IBSS_ATTR;
 long bpm IBSS_ATTR;
 
-long gmbuf[BUF_SIZE*NBUF];
+int32_t gmbuf[BUF_SIZE*NBUF];
 
 int quit=0;
 struct plugin_api * rb;
@@ -147,7 +147,7 @@ bool lastswap=1;
 
 inline void synthbuf(void)
 {
-   long *outptr;
+   int32_t *outptr;
    register int i;
    static int currentSample=0;
    int synthtemp[2];
